requirements.txt
streamlit
pydantic
google-generativeai
duckduckgo-search
anyio
aiohttp
tenacity
pyyaml
playwright
browser-use
